+++ projects/angular-phone-helper/src/lib/models/phone.model.ts
@@ -2,7 +2,8 @@ export interface IPhoneModel {
   code: string | null;
   codeDigit: string | null;
   format: string | null;
-  pattern: string | null;
+  pattern?: string | null;
+  formatInternational?: string | null;
 }
 
 export interface IPhonePipeOptions {