+++ projects/angular-phone-helper/src/lib/pipes/phone-format.pipe.ts
@@ -32,11 +32,10 @@ export class PhoneFormatPipe implements PipeTransform {
     const digitsOnly = removeNonDigits(phone) ?? '';
 
     // Find country by code, if it does not exist - chose international which is first
-    const countryHelper = getCountryHelperByCountryCode(countryIsoCode, this.config?.defaultCountryIsoCode);
+    const countryHelper = getCountryHelperByCountryCode(countryIsoCode, this.config?.defaultCountryIsoCode, this.config?.customCountry);
 
     // Check if country code is needed
     withCountryCode = checkIsItWithCountryCode(withCountryCode, this.config);
-
     // Apply the formatting pattern to the phone number
     let formattedPhone = convertToFormattedCountryPhone(digitsOnly, countryHelper, withCountryCode);